Testing Asynchronous Code
It's common in JavaScript for code to run asynchronously. When you have code that runs asynchronously, Jest needs to know when the code it is testing has completed, before it can move on to another test. Jest has several ways to handle this.

Promises

Return a promise from your test, and Jest will wait for that promise to resolve. If the promise is rejected, the test will fail.