.toContain(item) matcher
Use .toContain when you want to check that an item is in an array. For testing the items in the array, this uses ===, a strict equality check. .toContain can also check whether a string is a substring of another string.

This matcher also accepts others iterables such as strings, sets, node lists and HTML collections.