Asymmetric Matcher - expect.any(constructor)
expect.any(constructor) matches anything that was created with the given constructor or if it's a primitive that is of the passed type. You can use it inside toEqual or toHaveBeenCalledWith instead of a literal value. For example, if you want to check that a mock function is called with a number