Asymmetric Matcher - expect.arrayContaining(array)
expect.arrayContaining(array) matches a received array which contains all of the elements in the expected array. That is, the expected array is a subset of the received array. Therefore, it matches a received array which contains elements that are not in the expected array.

You can use it instead of a literal value:

• in toEqual or toHaveBeenCalledWith
• to match a property in objectContaining or toMatchObject