Asymetric Matchers - expect.stringMatching(string | regexp)
expect.stringMatching(string | regexp) matches the received value if it is a string that matches the expected string or regular expression.

You can use it instead of a literal value:

in toEqual or toHaveBeenCalledWith
to match an element in arrayContaining
to match a property in objectContaining or toMatchObject
This example also shows how you can nest multiple asymmetric matchers, with expect.stringMatching inside the expect.arrayContaining.