mockFn.mockReturnValueOnce(value)
Shorthand for:

jest.fn().mockImplementationOnce(() => value);

Accepts a value that will be returned for one call to the mock function. Can be chained so that successive calls to the mock function return different values. When there are no more mockReturnValueOnce values to use, calls will return a value specified by mockReturnValue.